Fix 2
Учёл предыдущие замечания

+++ 5383/Es'kov/arc.h
@@ -33,7 +33,7 @@ public:
 	void rotate(int angle);
 	void scale(float ratio);
 
-	float perimetr();
+	double perimetr();
 
 	friend std::ostream& operator<<(std::ostream& os, arc& a1);
 };
