+++ 5383/Es'kov/circle.h
@@ -25,7 +25,7 @@ public:
 	void rotate(int angle);
 	void scale(float ratio);
 
-	float perimetr();
+	double perimetr();
 
 	friend std::ostream& operator<<(std::ostream& os, circle& c1);
 };
